requirements.txt
gradio
python-dotenv
google-genai
openai
anthropic-sdk
mido
numpy
matplotlib
pillow
pydantic